pull out Web3 provider typings & mv from @truffle/blockchain-utils to @truffle/provider
